fix(works): added endpoint extension and reduced 'getImage' usage
